:recycle: Refactor database setup in tests using DbHelpers trait
